+++ core/components/filterwhere/lexicon/en/setting.inc.php
@@ -5,5 +5,10 @@
  * @package filterwhere
  * @subpackage lexicon
  */
+$_lang['area_geocode'] = 'Geocode';
 $_lang['setting_filterwhere.debug'] = 'Debug';
 $_lang['setting_filterwhere.debug_desc'] = 'Log debug information in the MODX error log.';
+$_lang['setting_filterwhere.google_maps_api_key'] = 'Google Maps Geocoding API Key';
+$_lang['setting_filterwhere.google_maps_api_key_desc'] = '<a href="https://developers.google.com/maps/documentation/javascript/get-api-key">Request</a> a Google Maps API Key with enabled Geocoding API.';
+$_lang['setting_filterwhere.google_maps_region'] = 'Google Maps API Region Code Biasing';
+$_lang['setting_filterwhere.google_maps_region_desc'] = '<a href="https://developers.google.com/maps/documentation/javascript/geocoding#GeocodingRegionCodes">Preferred region</a> for the geocoding result.';