+++ core/components/filterwhere/docs/changelog.md
@@ -2,8 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.1.0] - TBA
+
+### Added
+
+- RANGE operator
+- GEOCODE operator
+
+## [1.0.2] - 2023-01-25
+
+### Fixed
+
+- Fix PHP warning: strip_tags() expects parameter 1 to be string, array given
 
 ## [1.0.1] - 2021-11-28